+++ zsh/zfs.zsh
@@ -5,6 +5,7 @@
 # zfs
 alias z='zfs'
 alias zl='zfs list -o name,used,available,referenced,usedbysnapshots,compressratio,mountpoint'
+alias zl2='zfs list -o name,used,usedbysnapshots,readonly,compression,compressratio,quota,keystatus'
 alias zls='zfs list -r -d 1 -t snapshot -o name,used,refer,compressratio'
 alias zc='zfs create'
 alias zs='zfs snapshot'