+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * [Zsh/Docker] Add volume management aliases
 * [Zsh/ZFS] Add aliases for `sudo` operations
 * [Zsh/ZFS] Add `zl2` as a `zfs list` alias with different options than `zl`
+* [Zsh/ZFS] Add watch operations
 
 ## v0.0.5