+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 ### Enhancements
 
 * [Nix/Environment] Create the `$TMPDIR` before to rebuild in `nors`
-* [Nix/Utilities] Add `lshw`, `pciutils` and `usbutils`
+* [Nix/Utilities] Add `inxi`, `lshw`, `pciutils` and `usbutils`
 * [ranger/bepo] Add `È` to edit as root
 * [ranger/bepo] Add `gt` to go to `/tmp`
 * [Tmux] Use the color 231 instead of 15 for white to avoid issue with terminal