+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Breaking changes
 
+* [ranger/bepo] Open the current directory with Emacs on `e`
 * [Zsh/direnv] Make `dl` look at `~/Informatique` instead of `~/Programmes` and
     add a comment to make clear it is provided as an example