+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * [ranger/bepo] Add `gt` to go to `/tmp`
 * [Tmux] Use the color 231 instead of 15 for white to avoid issue with terminal
     theme switches
+* [Zsh/Aliases] Add `wnet` for `watch -n 1 nmcli`
 * [Zsh/Docker] Add start / stop aliases
 * [Zsh/Docker] Add `--sig-proxy=false` to the `docker attach` alias
 * [Zsh/Docker] Add volume management aliases