+++ CHANGELOG.md
@@ -20,8 +20,8 @@
 
 * [Nix/Environment] Create the `$TMPDIR` before to rebuild in `nors`
 * [Nix/Utilities] Add `colordiff`, `dnsutils`, `file`, `ffmpeg`, `inxi`, `jq`,
-    `lshw`, `parted`, `pciutils`, `qpdf`, `smartmontools`, `trickle` and
-    `usbutils`
+    `lshw`, `parted`, `pciutils`, `qpdf`, `smartmontools`, `telnet`,
+    `traceroute` and `usbutils`
 * [ranger/bepo] Add `È` to edit as root
 * [ranger/bepo] Add `gt` to go to `/tmp`
 * [Tmux] Use the color 231 instead of 15 for white to avoid issue with terminal